Fix reading not existing savedata
